pyproject.toml
[project]
name = "ai-agent-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "bcrypt==4.1.0",
    "duckduckgo-search>=8.1.1",
    "fastapi>=0.135.1",
    "httpx>=0.28.1",
    "langchain-ollama>=1.0.1",
    "langgraph>=1.0.10",
    "pydantic-settings>=2.13.1",
    "pydantic[email]>=2.12.5",
    "pytest>=9.0.2",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.22",
    "redis>=7.2.1",
    "slowapi>=0.1.9",
    "sqlalchemy>=2.0.48",
    "uvicorn>=0.41.0",
]
